Refatorando validação de usuário e login
